backend/requirements.txt
# NirovaAI Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0
slowapi==0.1.9

# Database
motor==3.4.0
pymongo==4.7.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# Cache
redis==5.0.4

# AI — LLM
groq==0.9.0
google-generativeai==0.7.2
langchain==0.2.5
langchain-community==0.2.5

# AI — Embeddings (runs locally, no API needed)
sentence-transformers==3.0.1

# AI — ML Models
xgboost==2.0.3
scikit-learn==1.5.0
onnxruntime==1.18.0
joblib==1.4.2

# AI — Vision
opencv-python-headless==4.10.0.82
Pillow==10.3.0

# Document Processing
PyMuPDF==1.24.5

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
certifi==2025.1.31
numpy==1.26.4
pandas==2.2.2
